html/editing/editing-0/making-entire-documents-editable-the-designmode-idl-attribute/user-interaction-editing-designMode-svg.svg

<svg:svg xmlns="http://www.w3.org/1999/xhtml" xmlns:svg="http://www.w3.org/2000/svg" width="100%" height="100%" viewBox="0 0 800 600"><svg:title>Editing: designMode attribute test</svg:title><head><link href="mailto: guopengcheng@baidu.com" rel="author" title="Baidu"/><link href="https://html.spec.whatwg.org/multipage/#making-entire-documents-editable:-the-designmode-idl-attribute" rel="help"/><script src="/resources/testharness.js"/><script src="/resources/testharnessreport.js"/><div id="log"/></head><body><script type="text/javascript"><![CDATA[
   test(function() {
    assert_equals(document.designMode, "off", "check for designMode value");
    assert_throws("InvalidStateError", function() { document.queryCommandSupported("delete") });
    assert_throws("InvalidStateError", function() { document.queryCommandEnabled("delete") });
   }, "initial designMode attribute");
   document.designMode="on";
   test(function() {
    assert_equals(document.designMode, "on", "check for designMode value");
    assert_throws("InvalidStateError", function() { document.queryCommandSupported("delete") });
    assert_throws("InvalidStateError", function() { document.queryCommandEnabled("delete") });
   }, "set designMode = \"on\"");
   document.designMode="off";
   test(function() {
    assert_equals(document.designMode,"off", "check for designMode value");
    assert_throws("InvalidStateError", function() { document.queryCommandSupported("delete") });
    assert_throws("InvalidStateError", function() { document.queryCommandEnabled("delete") });
   }, "set designMode = \"off\"");
  ]]></script></body></svg:svg>
